+++ lib/src/icustomanimationframerequester.dart
@@ -1,0 +1,9 @@
+part of babylon;
+
+/// interface ICustomAnimationFrameRequester
+@JS()
+abstract class ICustomAnimationFrameRequester {
+  Function renderFunction;
+  Function requestAnimationFrame;
+  num requestID;
+}