+++ lib/src/idisplaychangedeventargs.dart
@@ -1,0 +1,8 @@
+part of babylon;
+
+/// interface IDisplayChangedEventArgs
+@JS()
+abstract class IDisplayChangedEventArgs {
+  dynamic vrDisplay;
+  bool vrSupported;
+}