+++ lib/src/instancingattributeinfo.dart
@@ -1,0 +1,13 @@
+part of babylon;
+
+/// interface InstancingAttributeInfo
+@JS()
+abstract class InstancingAttributeInfo {
+  String attributeName;
+  num index;
+  num attributeSize;
+  num offset;
+  num divisor;
+  num attributeType;
+  bool normalized;
+}