+++ lib/src/ipipelinecontext.dart
@@ -1,0 +1,8 @@
+part of babylon;
+
+/// interface IPipelineContext
+@JS()
+abstract class IPipelineContext {
+  bool isAsync;
+  bool isReady;
+}