+++ lib/src/iscenelike.dart
@@ -1,0 +1,7 @@
+part of babylon;
+
+/// interface ISceneLike
+@JS()
+abstract class ISceneLike {
+  IOfflineProvider offlineProvider;
+}