+++ lib/src/iviewportlike.dart
@@ -1,0 +1,10 @@
+part of babylon;
+
+/// interface IViewportLike
+@JS()
+abstract class IViewportLike {
+  num x;
+  num y;
+  num width;
+  num height;
+}