+++ lib/src/iviewportownerlike.dart
@@ -1,0 +1,7 @@
+part of babylon;
+
+/// interface IViewportOwnerLike
+@JS()
+abstract class IViewportOwnerLike {
+  IViewportLike viewport;
+}