+++ lib/src/ivrpresentationattributes.dart
@@ -1,0 +1,8 @@
+part of babylon;
+
+/// interface IVRPresentationAttributes
+@JS()
+abstract class IVRPresentationAttributes {
+  bool highRefreshRate;
+  num foveationLevel;
+}